server/requirements.txt
openenv-core[core]>=0.2.3
openai>=2.7.2
